requirements.txt
fastapi
uvicorn
langchain
langchain_openai
python-dotenv
pydantic
pytest
httpx
wikipedia
langchain-community